Pull out interpolation into dedicated file
