+++ src/cli/color_swatch.rs
@@ -3,11 +3,7 @@ use serde::{Deserialize, Serialize};
 use crate::core::{
     color_map::{
         with_uniform_spacing, ColorMap, ColorMapKeyFrame, ColorMapLookUpTable, ColorMapper,
-        LinearInterpolator, StepInterpolator,
-    },
-    file_io::{serialize_to_json_or_panic, FilePrefix},
-    image_utils::write_image_to_file_or_panic,
-    stopwatch::Stopwatch,
+    }, file_io::{serialize_to_json_or_panic, FilePrefix}, image_utils::write_image_to_file_or_panic, interpolation::{LinearInterpolator, StepInterpolator}, stopwatch::Stopwatch
 };
 
 #[derive(Serialize, Deserialize, Debug)]
