+++ src/core/mod.rs
@@ -5,6 +5,7 @@ pub mod dynamical_systems;
 pub mod file_io;
 pub mod histogram;
 pub mod image_utils;
+pub mod interpolation;
 pub mod lookup_table;
 pub mod ode_solvers;
 pub mod render_window;
